최종발표 산출물/web/config/requirements.txt
langchain
langchain-community
openai
chromadb
tiktoken
pymysql
joblib
rapidfuzz
pandas
python-dotenv
requests
scikit-learn
chromadb
django
mysqlclient
django-cors-headers
whitenoise